requirements.txt
fastapi==0.136.0
imgkit==1.2.3
protobuf==7.34.1
python-dotenv==1.2.2
Requests==2.33.1
streamlit==1.56.0
uvicorn==0.44.0
google-genai
